+++ src/EntwineLLM/Helpers/PromptHelper.cs
@@ -4,24 +4,31 @@ using Newtonsoft.Json;
 
 namespace EntwineLlm.Helpers
 {
-    internal static class PromptHelper
+    internal class PromptHelper
     {
-        public static string CreateForManualRequest(string model, string userCode, string prompt)
+        private readonly string _language;
+
+        public PromptHelper(string language)
+        {
+            _language = language;
+        }
+
+        public string CreateForManualRequest(string model, string userCode, string prompt)
             => CreatePrompt(Properties.Resources.PromptForManual, model, userCode, prompt);
 
-        public static string CreateForRefactor(string model, string userCode)
+        public string CreateForRefactor(string model, string userCode)
             => CreatePrompt(Properties.Resources.PromptForRefactor, model, userCode);
 
-        public static string CreateForTests(string model, string userCode)
+        public string CreateForTests(string model, string userCode)
             => CreatePrompt(Properties.Resources.PromptForTests, model, userCode);
 
-        public static string CreateForDocumentation(string model, string userCode)
+        public string CreateForDocumentation(string model, string userCode)
             => CreatePrompt(Properties.Resources.PromptForDocumentation, model, userCode);
 
-        public static string CreateForReview(string model, string userCode)
+        public string CreateForReview(string model, string userCode)
             => CreatePrompt(Properties.Resources.PromptForReview, model, userCode);
 
-        private static string CreatePrompt(string promptModel, string model, string userCode, string manualRequest = "")
+        private string CreatePrompt(string promptModel, string model, string userCode, string manualRequest = "")
         {
             var promptText = PreparePrompt(promptModel);
             return ReplacePlaceholders(promptText, model, userCode, manualRequest);
@@ -34,12 +41,14 @@ namespace EntwineLlm.Helpers
                 .ReduceMultipleSpaces();
         }
 
-        private static string ReplacePlaceholders(string prompt, string model, string code, string manualRequest = "")
+        private string ReplacePlaceholders(string prompt, string model, string code, string manualRequest = "")
         {
             code = code.EscapeJsonString();
             manualRequest = manualRequest.EscapeJsonString();
 
-            var llmRequest = LlmRequest.Create(model, prompt, code, manualRequest);
+            var languagePrompt = Properties.Resources.SetLanguagePrompt.Replace("{LANGUAGE}", _language);
+
+            var llmRequest = LlmRequest.Create(model, languagePrompt, prompt, code, manualRequest);
             return JsonConvert.SerializeObject(llmRequest, Formatting.Indented, new JsonSerializerSettings()
             {
                 MissingMemberHandling = MissingMemberHandling.Ignore,