+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### v1.10.0 Added answer language
+---
+- It is now possibile to set a language in which the LLM must answer
+- Refactoring for LLM httpClient
+
 ### v1.9.0 Generic refactoring
 ---
 - Main menu moved to Extensions menu; command are now grouped